requirements.txt
torch
torchvision
gradio
Pillow
opencv-python
